+++ tests/functional/pytorch/jit/012_jit_resnet_path_tuning_one_sample.py
@@ -24,25 +24,16 @@ from logging import INFO, basicConfig
 import timm
 import torch
 
-from aitune.torch.jit.config import config
 from aitune.torch.jit.patched_module import PRINT_HIERARCHY_HEADER, PatchedModule
 
 
 def test_jit_resnet():
     import aitune.torch.jit.enable  # noqa: F401
 
-    config.min_samples = 2
-
     resnet = timm.create_model("resnet18", pretrained=False).to("cuda")
 
-    def batch():
-        # we are calling two times with different batch sizes to recognize dynamic axes
-        resnet(torch.randn(2, 3, 224, 224, device="cuda"))
-        resnet(torch.randn(16, 3, 224, 224, device="cuda"))
-
     with torch.no_grad():
-        for _ in range(5):
-            batch()
+        resnet(torch.randn(2, 3, 224, 224, device="cuda"))
 
     # Capture the print_hierarchy output
     history = []
@@ -53,8 +44,7 @@ def test_jit_resnet():
     assert PRINT_HIERARCHY_HEADER in history[0]
     assert re.match(r".*ResNet.*state=tuned.*TensorRTBackend", history[1])
 
-    assert resnet(torch.randn(8, 3, 224, 224, device="cuda")).shape == (8, 1000)
-    assert resnet(torch.randn(16, 3, 224, 224, device="cuda")).shape == (16, 1000)
+    assert resnet(torch.randn(2, 3, 224, 224, device="cuda")).shape == (2, 1000)
 
 
 if __name__ == "__main__":